GraphRAG-Based_Data_Search_Pipeline_3/requirements.txt
langchain
langchain_community
neo4j-graphrag
faiss-cpu
python-dotenv   
google-generativeai
sentence_transformers
pandas
matplotlib
seaborn